Set up basic project structure
Each day's puzzle gets its own subdirectory containing a
`solution.cpp/.h` pair and a `test.cpp` file. The solutions contain the
functions used to solve the respective puzzle. They are linked into a
static library which in turn is linked against by both the `aoc21`
executable (which runs it on the downloaded puzzle input on it to
produce the users' individualized solutions) as well as the
`aoc21-test` executable which contains the unit tests.